requirements.txt
Flask
Flask-Cors
requests
dnspython
